Laser Piano
It is just for our project which is inspired by the laser keybords.
Sounds so scentific! It should be used with camera and some special
filter to your camera. Specifically, it looks like a virtual laser
piano, which, in the software, needs control the hardware camera——needs
a filter to give up all the lights except infrared ray——to get actually
single input(of course the algorithm needs to do this work including
elimination of noise) and control the computer amplifing sounds like
piano when you play the "piano" in front of the camera. it is really a
simple one that we can just use it in the computer, but it includes the
corresponding coordinary( your hands play the real "piano", and the
camera inputs a pair of coordinary to match our database with working
out sounds), binarization and liminal value（it ebbs the noise to
accelerate progress), graph analysis ( we need to get the signal and
pick up the information which we need like the reflect light in your
fingers and transform it as the coordinary) , etc. other details you can
ask me and we can discuss through email. I feel very welcome to chat
with anybody! Come on!